refactor(components): refactor drawer (#4282)
