+++ packages/components/drawer/__tests__/drawer.spec.ts
@@ -1,7 +1,7 @@
 import { nextTick } from 'vue'
 import { mount } from '@vue/test-utils'
 import { rAF } from '@element-plus/test-utils/tick'
-import Drawer from '../src/index.vue'
+import Drawer from '../src/drawer.vue'
 import Button from '../../button/src/button.vue'
 
 jest.useFakeTimers()
