+++ packages/components/drawer/src/drawer.ts
@@ -1,0 +1,28 @@
+import { buildProps } from '@element-plus/utils/props'
+import { dialogProps, dialogEmits } from '@element-plus/components/dialog'
+import type { ExtractPropTypes } from 'vue'
+
+export const drawerProps = buildProps({
+  ...dialogProps,
+  direction: {
+    type: String,
+    default: 'rtl',
+    values: ['ltr', 'rtl', 'ttb', 'btt'],
+  },
+  size: {
+    type: [String, Number],
+    default: '30%',
+  },
+  withHeader: {
+    type: Boolean,
+    default: true,
+  },
+  modalFade: {
+    type: Boolean,
+    default: true,
+  },
+} as const)
+
+export type DrawerProps = ExtractPropTypes<typeof drawerProps>
+
+export const drawerEmits = dialogEmits
